+++ application/composer.json
@@ -3,5 +3,10 @@
         "google/cloud-translate": "^1.2",
         "kenjis/codeigniter-ss-twig": "^0.4.0",
         "illuminate/database": "^6.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "AdventistCommons\\Eloquent\\": "eloquent/"
+        }
     }
 }
