+++ application/config/autoload.php
@@ -140,9 +140,9 @@ $autoload['libraries'] = array();
 | -------------------------------------------------------------------
 |  Auto-load Eloquent ORM folder
 | -------------------------------------------------------------------
- */
 spl_autoload_register(function ($class) {
     if (file_exists(APPPATH . 'eloquent/' . $class . '.php')) {
         include APPPATH . 'eloquent/' . $class . '.php';
     }
 });      
+ */
