+++ application/eloquent/ProductContent.php
@@ -1,5 +1,5 @@
 <?php 
-
+namespace AdventistCommons\Eloquent;
 use \Illuminate\Database\Eloquent\Model as Eloquent;
 
 class ProductContent extends Eloquent {
