+++ application/eloquent/ProductSection.php
@@ -1,5 +1,5 @@
 <?php 
-
+namespace AdventistCommons\Eloquent;
 use \Illuminate\Database\Eloquent\Model as Eloquent;
 
 class ProductSection extends Eloquent {
